// web/src/api/models/question-set.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface QuestionSet
 */
export interface QuestionSet {
    /**
     * 
     * @type {string}
     * @memberof QuestionSet
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionSet
     */
    'user_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionSet
     */
    'collection_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionSet
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionSet
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionSet
     */
    'gmt_created'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionSet
     */
    'gmt_updated'?: string;
}